Backend/requirements.txt
flask
requests
flask-cors
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
python-dotenv
psycopg2-binary
